# backend/src/public/swagger.yaml
openapi: 3.0.0
info:
  title: Forkcast API
  description: API Documentation for Forkcast Application
  version: 1.0.0
servers:
  - url: http://localhost:3000/api
    description: Local development server
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  schemas:
...
    CalendarEntry:
       type: object
       properties:
         id:
           type: integer
         date:
           type: string
           format: date-time
         userId:
           type: integer
         recipeId:
           type: integer
    CalendarEntryWithRecipe:
       type: object
       properties:
         id:
           type: integer
         date:
           type: string
           format: date-time
         userId:
           type: integer
         recipeId:
           type: integer
         recipe:
           allOf:
             - $ref: '#/components/schemas/RecipePreview'
             - type: object
               nullable: true
    User:
       type: object
       required:
         - id
         - name
         - email
         - profilePicture
       properties:
        id:
          type: integer
        name:
          type: string
        email:
          type: string
          format: email
        profilePicture:
          type: string
          format: uri
          nullable: true
    LoginResponse:
      type: object
      required:
        - user
        - token
      properties:
        user:
          $ref: '#/components/schemas/User'
        token:
          type: string
          description: JWT Bearer token
    RegisterRequest:
      type: object
      required:
        - name
        - email
        - password
      properties:
        name:
          type: string
        email:
          type: string
          format: email
        password:
          type: string
          format: password
          description: Must satisfy the backend strong-password validation.
    LoginRequest:
      type: object
      required:
        - identifier
        - password
      properties:
        identifier:
          type: string
          description: Email or Username
        password:
          type: string
          format: password
    VerifyRequest:
      type: object
      required:
        - email
        - code
      properties:
        email:
          type: string
          format: email
        code:
          type: string
          pattern: '^\d{6}$'
          description: 6-digit verification code
    ForgotPasswordRequest:
      type: object
      required:
        - email
      properties:
        email:
          type: string
          format: email
          description: The email address to send the reset code to
    ResetPasswordRequest:
      type: object
      required:
        - email
        - code
        - password
      properties:
        email:
          type: string
          format: email
        code:
          type: string
          pattern: '^\d{6}$'
          description: 6-digit verification code
        password:
          type: string
          format: password
          description: Must satisfy the backend strong-password validation.
...
    AddCalendarEntryRequest:
      type: object
      required:
        - recipeId
        - date
      properties:
        recipeId:
          type: integer
        date:
          type: string
          format: date-time
          description: ISO 8601 date string.
...
    ValidationErrorItem:
      type: object
      properties:
        type:
          type: string
        value:
          nullable: true
        msg:
          type: string
        path:
          type: string
        location:
          type: string
    ValidationErrorResponse:
      type: object
      properties:
        errors:
          type: array
          items:
            $ref: '#/components/schemas/ValidationErrorItem'
    MessageResponse:
      type: object
      properties:
        message:
          type: string
...
    RecipePreviewRating:
      type: object
      required: [rating, count]
      properties:
        rating:
          type: number
          minimum: 1
          maximum: 5
          description: >-
            Average rating. Reflects the live user-rating average (1 decimal) when at least
            one user has rated the recipe; otherwise falls back to the Spoonacular score.
        count:
          type: integer
          description: >-
            Number of user ratings when any exist; otherwise the Spoonacular aggregate-likes count.
    RecipePreviewAttribute:
      type: object
      required: [icon, text]
      properties:
        icon:
          type: string
          description: Font Awesome icon name (without fa- prefix)
        text:
          type: string
    RecipePreviewTag:
      type: object
      required: [icon, text, color]
      properties:
        icon:
          type: string
          description: Font Awesome icon name (without fa- prefix)
        text:
          type: string
        color:
          type: string
          description: DaisyUI badge color (success, warning, primary, error, neutral)
    RecipePreview:
      type: object
      required: [id, title, image, effort, rating, attributes, tags]
      properties:
        id:
          type: integer
        title:
          type: string
        image:
          type: string
          format: uri
        effort:
          type: integer
          minimum: 1
          maximum: 100
          description: Effort score calculated from cook time, step count and ingredient count
        rating:
          $ref: '#/components/schemas/RecipePreviewRating'
        attributes:
          type: array
          items:
            $ref: '#/components/schemas/RecipePreviewAttribute'
          description: Display attributes (e.g. cook time, servings)
        tags:
          type: array
          items:
            $ref: '#/components/schemas/RecipePreviewTag'
          description: Dietary tags (Vegan, Vegetarian, Gluten Free, Dairy Free)
        isFavorited:
          type: boolean
          description: Whether the authenticated user has favorited this recipe. Absent when the caller is unauthenticated.
        userRating:
          type: integer
          minimum: 1
          maximum: 5
          description: The authenticated user's own rating for this recipe. Absent when the caller is unauthenticated or has not rated it.
...
security:
  - bearerAuth: []
paths:
  /auth/register:
    post:
      summary: Register a new user
      tags: [Auth]
      security: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/RegisterRequest'
      responses:
        '201':
          description: User registered successfully (Pending Verification)
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/User'
        '400':
          description: Invalid input
          content:
            application/json:
              schema:
                  $ref: '#/components/schemas/ValidationErrorResponse'
        '500':
          description: Internal server error
  /auth/login:
    post:
      summary: Login a user and return a JWT token
      tags: [Auth]
      security: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/LoginRequest'
      responses:
        '200':
          description: Login successful
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/LoginResponse'
        '401':
          description: Invalid credentials
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/MessageResponse'
        '403':
          description: Account not verified
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/MessageResponse'
        '500':
          description: Internal server error
  /auth/verify:
    post:
      summary: Verify email with code
      tags: [Auth]
      security: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/VerifyRequest'
      responses:
        '200':
          description: Code verified successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/MessageResponse'
        '400':
          description: Invalid input
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ValidationErrorResponse'
        '401':
          description: Invalid code
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/MessageResponse'
        '500':
          description: Internal server error
  /auth/password/forgot:
    post:
      summary: Request a password reset code
      tags: [Auth]
      security: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/ForgotPasswordRequest'
      responses:
        '200':
          description: Code sent successfully (if user exists)
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/MessageResponse'
        '400':
          description: Invalid input
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ValidationErrorResponse'
        '500':
          description: Internal server error
  /auth/password/reset:
    post:
      summary: Reset the user's password using the code
      tags: [Auth]
      security: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/ResetPasswordRequest'
      responses:
        '200':
          description: Password reset successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/MessageResponse'
        '400':
          description: Invalid input
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ValidationErrorResponse'
        '401':
          description: Invalid code
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/MessageResponse'
        '404':
          description: User not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/MessageResponse'
          '500':
            description: Internal server error
...
  /users/me/calendar:
    get:
      summary: Get calendar entries
      tags: [Calendar]
      parameters:
        - in: query
          name: from
          schema:
            type: string
            format: date
          description: Start date in ISO8601 format (YYYY-MM-DD). Optional, returns all if omitted.
        - in: query
          name: to
          schema:
            type: string
            format: date
          description: End date in ISO8601 format (YYYY-MM-DD). Optional, returns all if omitted.
        - in: query
          name: populate
          schema:
            type: boolean
          description: 'When true, embeds full recipe details (including isFavorited) inside each entry. When false (default), returns only raw CalendarEntry objects.'
      responses:
        '200':
          description: Calendar entries (with or without recipe details based on populate flag)
          content:
            application/json:
              schema:
                type: array
                items:
                  oneOf:
                    - $ref: '#/components/schemas/CalendarEntry'
                    - $ref: '#/components/schemas/CalendarEntryWithRecipe'
        '400':
          description: Invalid query parameters
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ValidationErrorResponse'
        '401':
          description: Missing or invalid authentication
        '500':
          description: Internal server error
    post:
      summary: Add calendar entry
      tags: [Calendar]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/AddCalendarEntryRequest'
      responses:
        '200':
          description: Calendar entry added
        '400':
           description: Invalid request body
           content:
            application/json:
              schema:
                $ref: '#/components/schemas/ValidationErrorResponse'
        '401':
          description: Missing or invalid authentication
        '500':
          description: Internal server error
...
  /users/me/calendar/{calendarEntryId}:
    delete:
      summary: Remove calendar entry
      tags: [Calendar]
      parameters:
        - in: path
          name: calendarEntryId
          required: true
          schema:
            type: integer
      responses:
        '200':
          description: Removed calendar entry
        '400':
           description: Invalid ID
        '401':
          description: Missing or invalid authentication
        '404':
          description: Not found
        '500':
          description: Internal server error
    get:
      summary: Get calendar entry
      tags: [Calendar]
      parameters:
        - in: path
          name: calendarEntryId
          required: true
          schema:
            type: integer
      responses:
        '200':
          description: Calendar entry details
        '404':
          description: Not found
        '401':
           description: Missing or invalid authentication
        '500':
           description: Internal server error